Removed trending and random buttons
Had to revert to the previous version of code since there were errors blocking the regular functions of the gif search. When doing a regular search, typeahead would work and load the gifs but when you pressed enter, it would direct to the trending gifs instead. This is most likely a route error and I'll have to comb through and fix it. But for now, in order to submit functioning code, I'll remove it then fix it afterward.